+++ lib/given_when_then_unit_test.dart
@@ -1,3 +1,6 @@
 library given_when_then_unit_test;
 
-export './res/res.dart';
+export 'res/given.dart';
+export 'res/test_fixtures.dart';
+export 'res/then.dart';
+export 'res/when.dart';