+++ CHANGELOG.md
@@ -1,4 +1,10 @@
+## 0.0.2
+
+* [Add] [examples](./example/lib/example.dart)
+* Refactoring
+* 100% test coverage
+
 ## 0.0.1
 
-* Add `given` `when` `then` methods
-* Add `before`, `after`, `beforeAll`, `afterAll` fixtures
+* [Add] `given` `when` `then` methods
+* [Add] `before`, `after`, `beforeAll`, `afterAll` test fixtures